# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-Commerce API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  comparePrice?: number;
  stock?: number;
  imageUrl?: string;
  featured?: boolean;
  categoryId?: number;
}